requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
numpy
